Bug Fixes, Player UI Class, Board Class.
Finally got rid of the warnings, they got a bit annoying. Added the PlayerUI class, Board class, in testing.Frame there is a little demo of both. Added the abstract class Space for when I implement the game's logic so I can call land(Player p) easier. The only thing I have left to hard-code is the game's logic and how the player deals with being in jail, then it's off to TONS of bug fixes because I'm sure there's a ton wrong just because I haven't been able to test it.